requirements.txt
# Core Framework
google-adk>=1.25.0
flask>=3.0.0
pydantic>=2.0
sqlalchemy>=2.0

# Networking
httpx>=0.27.0
requests>=2.31.0

# Graph & Analysis
networkx>=3.0

# Web Search (all free, no API keys)
ddgs>=9.0.0
trafilatura>=2.0.0
wikipedia>=1.4.0

# Configuration
python-dotenv>=1.0.0
click>=8.0

# Dev / Testing
pytest>=8.0
